Add Web3 Permissions, LocalStorage, Extensions.json->.ts, Add Tests
Here, we add Web3 permissions, specifically allowing only given ABI functions and selectors, e.g.

```json
"permissions": {
  "storage": "*",
  "trx": [
    {
      "to": "$operator",
      "abi": "migrate((address,uint256)[],uint256)",
      "params": "*"
    },
    {
      "to": "*",
      "abi": "approve(address,uint256)",
      "params": ["$operator", "*"]
    }
  ]
}
```

This means that the Extension is allowed to read/write to localStorage, and we allow it to call `migrate()` on its operator as well as call `approve($operator, *)` on any Erc20 token. This is a strong system to prevent phishing from third-party extensions.

Also, we add a LocalStorage shim that uses RPC messages.

Also, we change Extensions.json to Extensions.ts for better type-checking (and slightly prettier syntax with as many `"`s).

Also, we add test cases, starting with strong test cases for Permissions, since it's becoming complex already, with a variety of "*" allowances and now ABI parsing.