Add Sudo and Sync Async
This patch adds a few extra perms, including "sudo" that is meant to grant all permissions, which can be used, for instance, in testing and developing extensions. We also implement `provider.sendAsync`, which is the out-dated way of sending messages through a provider, but apparently is used by some older extensions.

We also add some permissions for `modals` and `popups`, which simply add extra sandbox attributes to iframes to allow opening new windows and/or creating "alerts".